fix: number type
real number is now of type double

+++ src/number.h
@@ -1,6 +1,6 @@
 #ifndef NUMBER_H
 #define NUMBER_H
 
-typedef int number;
+typedef double number;
 
 #endif // NUMBER_H
